requirements.txt
SpeechRecognition==3.10.1
openai-whisper==20231117
requests==2.31.0
beautifulsoup4==4.12.3
pyttsx3==2.90
PyAudio==0.2.14
python-dotenv==1.0.1
keyboard==0.13.5
loguru==0.7.2 